complete code refactoring for improved structure and maintainability
