+++ README.md
@@ -6,10 +6,13 @@ An intuitive application designed to create stunning color palettes, tailored fo
 
 ## Overview
 
-<img src="https://i.imgur.com/5QgPpRL.png" alt="table image of https://palette-generator-indol.vercel.app" width="200" />
-<img src="https://i.imgur.com/d1QLQ3C.png" alt="android image of https://palette-generator-indol.vercel.app" width="200" />
-<img src="https://i.imgur.com/omXHxoO.png" alt="desktop image of https://palette-generator-indol.vercel.app" width="200" />
-<img src="https://i.imgur.com/gqxpYwT.png" alt="iphone image of https://palette-generator-indol.vercel.app" width="200" />
+<img src="https://i.imgur.com/sY1cd9D.png" alt="table image of https://palette-generator-indol.vercel.app" width="200" />
+<img src="https://i.imgur.com/Sk1rZ8p.png" alt="desktop image of https://palette-generator-indol.vercel.app" width="200" />
+<img src="https://i.imgur.com/kXWEGWI.png" alt="iphone image of https://palette-generator-indol.vercel.app" width="200" />
+
+You can also install the app for better functionality.
+
+<img src="https://i.imgur.com/ufkfY7S.png" alt="pwa image of https://palette-generator-indol.vercel.app" width="250" >
 
 ## Usage2
