+++ src/App.tsx
@@ -8,15 +8,15 @@ import {
   useState
 } from "react";
 
-import { SavePalette } from "@components/SavePalette.tsx";
+import { SavePalette } from "@/components/save-palette/SavePalette";
 import { Toaster } from "@components/shared/ui/sonner";
 import { useLocation } from "wouter";
 import { ValidateHexadecimal } from "./utils/hexadecimal-validator";
 
 import chroma from "chroma-js";
-import GraphicItems from "@components/GraphicItems";
+import GraphicItems from "@components/graphic-items/GraphicItems";
 import Layout from "./layouts/Layout";
-import Palette from "@components/Palette";
+import Palette from "@/components/palette/Palette";
 import ColorPicker from "./components/color-picker/ColorPicker";
 import { HexadecimalContext } from "./provider/hexadecimal/hexadecimal.context";
 import SavePaletteSection from "./sections/SavePalette";
@@ -60,13 +60,13 @@ function App() {
 
   return (
     <Layout>
-      <section className="pt-24 font-sans">
+      <section className="pt-10 sm:pt-24 font-sans">
         <div
           style={{ "--color": deferredColor + "64" }}
           className="absolute inset-0 bg-gradient-to-b from-[var(--color)] to-white to-25% h-full -z-10"
         />
         <div className="flex flex-col items-center justify-center w-full h-full mx-auto gap-[36px] mb-40">
-          <h1 className="text-3xl font-bold font-headings lg:text-6xl">
+          <h1 className="text-3xl font-bold font-headings lg:text-6xl text-center md:mb-20">
             Generate your Custom Palette
           </h1>
           <Toaster />
