+++ src/components/copy-palette/CopyPalette.tsx
