+++ src/components/delete-palette/DeletePalette.tsx
