+++ src/components/palette/Palette.tsx
